Update patch tests
